ghetto fix for failing loop mount
This happens when mv_initrd_to_image.sh is invoked directly (ie not
via run.sh), and possibly when invoked by run.sh too if one or more of
the earlier steps are skipped.

This should be investigated further in order to find the root cause,
for a real fix.