move env ST_STBOOT_URL into new global config file run.config
We let sub-tasks (ie scripts invoked by run.sh) find run.config by
passing the full path to it in $1. This is for keeping sub-tasks as
independent as possible from run.sh and run.config.

If we later decide to promote run.sh to the almighty driver of all
things we could add knowledge about run.config to all the scripts that
need it.