stboot-installation/mbr-bootloader: Add fetch_syslinux.sh (#57)
This commit fixes the MBR bootloader build.

build_boot_filesystem.sh needs the syslinux fetch, which is is done
in build_image.sh. Therefor make_mbr_bootloader does not work on a
clean repository. This commit creates a separate script for fetching
syslinux and is called in make_mbr_bootloade.sh to call it before
bould_boot_filesystem.sh.

Signed-off-by: Marcello Sylvester Bauer <sylv@sylv.io>