Makefile to build targets (#65)
Create a Makefile to reflect the operation in "run.sh".

Use make to build the toolchain. It supports multithreaded build,
which accelerates the process.

Since many targets depends on specific config variables, this commit
adds an function structure to generate subconfig files for specific
targets. The function definition includes a comment with instructions
how to use it.

Add EFI application target analog to MBR bootloader.
Move common target from mbr-bootloader to common/makefile.

Use phony target instead of binary to force updates, even if binary
already exist.

An error will be printed if the file ST_SIGNING_ROOT does not exist.

The initramfs build depends on u-root, which is updated every
run to stay up-to-date. To prevent unnecessary rebuilds a checksum
file is generated as refence if the u-root binary changes.

Since the OS kernel and initramfs depends on the phony target acm,
it gets rebuild every time it is needed. To avoid this, OS packages
must be explicitly build if needed.

Signed-off-by: Marcello Sylvester Bauer <sylv@sylv.io>